feat(blueprint): add link to resource data mappings in a blueprint instance
Adds a new field to link state that holds a mapping of fields
in any resource in the same blueprint to the field path in the
link data.
This is primarily useful in that it allows the drift checker
to overlay link data in the same blueprint on resource state
so that updates made by links to any resource in the same blueprint
are not picked up as drift.
This is essential for the concept of links as side effects that can
update resources linked together along with existing resources
in the same blueprint that act as intermediaries, without this
behaviour, any changes that links make will be consistently picked up
as drift and if you relax the drift checker then it will be very
difficult to track the current state and the abstraction of links
would become an unwieldy addition that would make things less
predictable.
With the combination of the link data projection and link changes
included when staging changes, there is a clear picture of what
the current state should be looking at the resource spec data as
defined by the user and the updates made by linking resources
together. Drift detection now takes link changes into account,
based on a similar picture to what the user would see when staging
changes.