feat(blueprint): add support for path array selectors
Adds support for path array selectors for mapping paths used
for searching for values in mapping nodes along with injecting
values into mapping nodes.
This also includes a breaking change to provide a less confusing
name for the version of the inject into path function that supports
replacing object fields.
This also includes refactoring of the path parser to a more maintainable
and extensible top-down parser inspired by the substitutions parser.