fix(blueprint): add correction to mapping node injection behaviour
Adds fix to mapping node injection behaviour
that makes sure entire structures are not replaced
with an empty array or mapping when injecting in
"replace fields" mode. Arrays and mappings should only
be created if the value in the injection path does not
already exist in the structure that the value is being
injected into.