fix(blueprint): add final fix for empty schema validation checks
Adds final set of corrections to ensure nullable fields
in a schema are treated as expected, the previous
implementation would return an error for top-level mapping
node nil values. This also adds a correction to make sure
that the scalar type is inspected for nil values in addition
to the top level mapping node.