fix(blueprint): add missing suggested action for abstract resource type not found
An abstract resource type is retrieved as a fallback
when a concrete resource type in a provider could not be
found for a given resource type.
In the context that this error is used, the resource type
can exist as a part of a provider or a transformer
so we need to add a suggested action to check resource types
by provider.