feat(blueprint): add support for runtime binary data for functions
Adds support for holding binary data as slice of bytes
at runtime to allow functions in ${..} to process binary
data. This is particularly useful for tasks like reading
a file and then choosing whether base64 or utf-8 encode it
to a string, it can also be useful for base64 decoding a string
to be passed into another function that processes bytes.