fix(blueprint): remove file source registry from function call context
The file source registry should only be used by the file
function implementation, this commit simplifies the dependencies
so that the file function implementation takes the file source
registry on construction. This also prevents confusion around the
call context interface for provider plugin developers as the file
source registry would have been a part of the call context but
would not be callable.