feat(blueprint): add support for none values in substitutions
Adds support for special `none` values in substitutions as per
the latest iteration of the blueprint specification.
This also adds a fix to correctly prepare variadic args for function calls
in the substitution engine.
A new `coalesce` function has also been added that is distinct from
`first` in that it returns the first value that is not `none`, allowing
for other empty values such as empty strings.