feat(cli): add stage command for change staging
Adds the implementation of the `stage`
command for change staging. This includes an implementation
with a reusable split pane tui component housed in
the deploy cli sdk and enhancements to the e2e testing
infrastructure to include a stub provider in order to carry
out e2e tests for change staging.