fix(plugin-framework): allow empty resource spec states to reveal more useful errors
Allows empty mapping nodes for resource spec state
to reveal more meaningful errors for the end user
instead of serialisation error: missing mapping node.
Often, when checking drift for change staging,
a deployment may be in a partial or failed state
and it is better to reveal detailed errors such as missing
id required to get an external resource instead.