feat(blueprint-state): add persistence for reconciliation results
Adds persistence support for host applications such as the deploy engine
for reconciliation results.
Reconciliation results are their own entity exposed via the `manage` package
used by the deploy engine to manage entities outside of the scope of
the core blueprint library.
This also adds optimisations to include useful indexes and infrastructure
to facility time-based partitioning for the events table in the postgres
storage engine.